feat: use floats for numbers
This change introduces `double` as the type backing `number_t` as
opposed to `int32_t`. This is to keep the language higher level with
only one type of number, yet allow a wider array of use cases. This is
in line with what other LISPs and previous lifp implementation did.

We don't expect this to have substantial effects on memory usage, since
all the numbers are boxed.