feat: add return arena for better lifecycle management
With this change, evaluate will take two values: one for the values to
return, and the scratch area. This segregation is to make the two type
of lifecycle explicitly different: the former needs to be persisted and
can be used by the callers (e.g., the REPL), whereare the second is not
expected to be used by the callers.

Ths change does not incur an increase in memory usage.