refactor: drop scratch memory optimizations
As expected, arena frames were not the best way to handle transient values; they worked well in principle, but turned into an endless run of random values disappearing.

This change removes all the arena frames and directly maps lifecycles to arenas. It's a bit more expensive in terms of memory, but it's a lot more safe.

We will stop (for now!) chasing memory optimizations in favour of having a more predictable runtime