feat: handle closure with capture environments
With this change, the `fn` special form creates an environment in the
same arena where the closure is defined and populates it with values
from the above environments based `closure.form`.

This approach is consistent with arena handling (the closure and the
environment will have the same lifecycle), but it's wasteful as you will
create an environment you don't know you will use.

This change also has a bunch of TODOs to address in following commits,
it's mostly here to test viability of this solution.