refactor: make vm methods idiomatic and remove max call stack
With the new change, the only bound for lifp is the available memory you decide to run with in the beginning. Call Stack does not need to be artificially bound.

This change also cleans up the invokeClosure method, removing the unused environment parameter and adds error handling to environment capture.